Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -18,7 +18,7 @@ int main() {
                 printf("error\n");
             }
             else{
-                printf("%.0f\n",(float)a/b);
+                printf("%.0f\n",a/b);
             }
         break;
         default:
